+++ src/yb/util/decimal-test.cc
@@ -211,7 +211,7 @@ TEST_F(DecimalTest, TestComparableEncoding) {
   std::vector<Decimal> test_decimals;
   std::vector<std::string> encoded_strings;
   std::vector<Decimal> decoded_decimals;
-  for (int i = 0; i < test_cases.size(); i++) {
+  for (size_t i = 0; i < test_cases.size(); i++) {
     SCOPED_TRACE(Format("Index: $0, value: $1", i, test_cases[i]));
     test_decimals.emplace_back(test_cases[i]);
     encoded_strings.push_back(test_decimals[i].EncodeToComparable());
@@ -236,7 +236,7 @@ TEST_F(DecimalTest, TestBigDecimalEncoding) {
   std::vector<std::string> encoded_strings;
   std::vector<Decimal> decoded_decimals;
   bool is_out_of_range = false;
-  for (int i = 0; i < test_cases.size(); i++) {
+  for (size_t i = 0; i < test_cases.size(); i++) {
     SCOPED_TRACE(Format("Index: $0, value: $1", i, test_cases[i]));
     test_decimals.emplace_back(test_cases[i]);
     encoded_strings.push_back(test_decimals[i].EncodeToSerializedBigDecimal(&is_out_of_range));
